ref(github): Split issue tool into its own module
Move the GitHub createIssue implementation into a per-tool module and keep the root tools file focused on assembling plugin tools. This keeps the package ready for additional tools without growing one implementation file.\n\nRefs GH-660\nCo-Authored-By: GPT-5 Codex <codex@openai.com>
